sub_server/requirements.txt
anyio==3.7.1
idna==3.4
Pillow==10.0.0
sniffio==1.3.0
starlette==0.31.0
